requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
gTTS==2.2.4
idna==3.4
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
